grk_buffer_2d alloc: fix overflow bug in data size calculation

+++ src/lib/jp2/util/util.h
@@ -269,7 +269,7 @@ template <typename T> struct grk_buffer_2d : public grk_rect_u32 {
 	bool alloc(bool clear){
 		if (!data && width() && height()) {
 			stride = grk_make_aligned_width(width());
-			uint64_t data_size_needed = stride * height() * sizeof(T);
+			uint64_t data_size_needed = (uint64_t)stride * height() * sizeof(T);
 			if (!data_size_needed)
 			  return true;
 			data = (T*) grk_aligned_malloc(data_size_needed);
